Refactor type conversion utilities into shared module
