+++ dipeo/infrastructure/codegen/ir_builders/utils.py
@@ -6,279 +6,18 @@ import re
 from pathlib import Path
 from typing import Any, Optional
 
-import inflection
 import yaml
 
-# ============================================================================
-# CASE CONVERSION UTILITIES
-# ============================================================================
-
-
-def snake_case(text: str) -> str:
-    """Convert text to snake_case using inflection library."""
-    if not text or text == "Undefined":
-        return ""
-    return inflection.underscore(str(text))
-
-
-def camel_case(text: str) -> str:
-    """Convert text to camelCase using inflection library."""
-    if not text or text == "Undefined":
-        return ""
-    return inflection.camelize(str(text), uppercase_first_letter=False)
-
-
-def pascal_case(text: str) -> str:
-    """Convert text to PascalCase using inflection library."""
-    if not text or text == "Undefined":
-        return ""
-    return inflection.camelize(str(text))
-
-
-def kebab_case(text: str) -> str:
-    """Convert text to kebab-case using inflection library."""
-    if not text or text == "Undefined":
-        return ""
-    return inflection.dasherize(inflection.underscore(str(text)))
-
-
-# Aliases for backward compatibility
-camel_to_snake = snake_case
-snake_to_pascal = pascal_case
-pascal_to_camel = camel_case
-
-
-# ============================================================================
-# TYPE CONVERSION UTILITIES
-# ============================================================================
-
-
-class TypeConverter:
-    """Unified type converter for TypeScript, Python, and GraphQL types."""
-
-    # Base type mappings
-    TS_TO_PYTHON_BASE = {
-        "string": "str",
-        "number": "float",
-        "boolean": "bool",
-        "any": "Any",
-        "unknown": "Any",
-        "null": "None",
-        "undefined": "None",
-        "void": "None",
-        "Date": "datetime",
-        "Record<string, any>": "Dict[str, Any]",
-        "Record<string, string>": "Dict[str, str]",
-        "object": "Dict[str, Any]",
-    }
-
-    TS_TO_GRAPHQL = {
-        "string": "String",
-        "number": "Float",
-        "boolean": "Boolean",
-        "any": "JSONScalar",
-        "unknown": "JSONScalar",
-        "Date": "DateTime",
-        "Record<string, any>": "JSONScalar",
-        "Record<string, string>": "JSONScalar",
-        "object": "JSONScalar",
-    }
-
-    GRAPHQL_TO_TS = {
-        "String": "string",
-        "Int": "number",
-        "Float": "number",
-        "Boolean": "boolean",
-        "ID": "string",
-        "DateTime": "string",
-        "JSON": "any",
-        "JSONScalar": "any",
-        "Upload": "File",
-    }
-
-    def __init__(self, custom_mappings: Optional[dict[str, dict[str, str]]] = None):
-        """Initialize with optional custom type mappings.
-
-        Args:
-            custom_mappings: Dictionary with keys like 'ts_to_python', 'ts_to_graphql', etc.
-        """
-        self.custom_mappings = custom_mappings or {}
-
-    def ts_to_python(self, ts_type: str) -> str:
-        """Convert TypeScript type to Python type."""
-        # Check custom mappings first
-        if "ts_to_python" in self.custom_mappings:
-            if ts_type in self.custom_mappings["ts_to_python"]:
-                return self.custom_mappings["ts_to_python"][ts_type]
-
-        # Handle known type aliases to avoid forward reference issues
-        if ts_type == "SerializedNodeOutput":
-            return "SerializedEnvelope"
-        if ts_type == "PersonMemoryMessage":
-            return "Message"
-
-        # Handle union types (A | B | C) - check this before string literals
-        # to properly handle unions of string literals
-        if "|" in ts_type:
-            return self._handle_union_type(ts_type, self.ts_to_python)
-
-        # Handle string literals (only after checking for unions)
-        if self._is_string_literal(ts_type):
-            return f"Literal[{ts_type}]"
-
-        # Handle array types
-        if ts_type.startswith("Array<") and ts_type.endswith(">"):
-            inner_type = ts_type[6:-1]
-            return f"List[{self.ts_to_python(inner_type)}]"
-
-        if ts_type.endswith("[]"):
-            inner_type = ts_type[:-2]
-            return f"List[{self.ts_to_python(inner_type)}]"
-
-        # Handle Record types generically
-        if ts_type.startswith("Record<") and ts_type.endswith(">"):
-            # Extract key and value types from Record<K, V>
-            inner = ts_type[7:-1]  # Remove "Record<" and ">"
-            parts = inner.split(",", 1)
-            if len(parts) == 2:
-                key_type = parts[0].strip()
-                value_type = parts[1].strip()
-                # Convert both types recursively
-                py_key = self.ts_to_python(key_type)
-                py_value = self.ts_to_python(value_type)
-                # Map number to int for dictionary keys
-                if py_key == "float":
-                    py_key = "int"
-                return f"Dict[{py_key}, {py_value}]"
-
-        # Handle TypeScript object types: { ... }
-        if ts_type.strip().startswith("{") and ts_type.strip().endswith("}"):
-            # Check if it's an index signature type like { [key: string]: string; }
-            import re
-
-            pattern = r"\[\s*\w+\s*:\s*(\w+)\s*\]\s*:\s*(\w+)"
-            match = re.search(pattern, ts_type)
-            if match:
-                key_type = match.group(1)
-                value_type = match.group(2)
-                py_key = self.ts_to_python(key_type)
-                py_value = self.ts_to_python(value_type)
-                # Map number to int for dictionary keys
-                if py_key == "float":
-                    py_key = "int"
-                return f"Dict[{py_key}, {py_value}]"
-
-            # Check if it's an inline object type with specific properties
-            # e.g., { totalFiles: number; successCount: number; }
-            # For now, convert to Dict[str, Any] for simplicity
-            # TODO: Could generate a proper model class for these
-            if ":" in ts_type and ";" in ts_type:
-                return "Dict[str, Any]"
-
-        # Handle branded scalars
-        if "&" in ts_type and "__brand" in ts_type:
-            match = re.search(r"'__brand':\s*'([^']+)'", ts_type)
-            if match:
-                return match.group(1)
-
-        # Use base mapping
-        return self.TS_TO_PYTHON_BASE.get(ts_type, ts_type)
-
-    def ts_to_graphql(self, ts_type: str) -> str:
-        """Convert TypeScript type to GraphQL type."""
-        # Check custom mappings first
-        if "ts_to_graphql" in self.custom_mappings:
-            if ts_type in self.custom_mappings["ts_to_graphql"]:
-                return self.custom_mappings["ts_to_graphql"][ts_type]
-
-        # Handle arrays
-        if ts_type.startswith("Array<") and ts_type.endswith(">"):
-            inner_type = ts_type[6:-1]
-            return f"[{self.ts_to_graphql(inner_type)}]"
-
-        if ts_type.endswith("[]"):
-            inner_type = ts_type[:-2]
-            return f"[{self.ts_to_graphql(inner_type)}]"
-
-        # Handle branded scalars
-        if "&" in ts_type and "__brand" in ts_type:
-            match = re.search(r"'__brand':\s*'([^']+)'", ts_type)
-            if match:
-                # Convert to GraphQL scalar name
-                brand = match.group(1)
-                return brand if brand.endswith("ID") else brand
-
-        # Use base mapping
-        return self.TS_TO_GRAPHQL.get(ts_type, ts_type)
-
-    def graphql_to_ts(self, graphql_type: str) -> str:
-        """Convert GraphQL type to TypeScript type."""
-        # Check custom mappings first
-        if "graphql_to_ts" in self.custom_mappings:
-            if graphql_type in self.custom_mappings["graphql_to_ts"]:
-                return self.custom_mappings["graphql_to_ts"][graphql_type]
-
-        # Handle arrays
-        if graphql_type.startswith("[") and graphql_type.endswith("]"):
-            inner = graphql_type[1:-1].replace("!", "")
-            return f"{self.graphql_to_ts(inner)}[]"
-
-        # Remove required marker
-        clean_type = graphql_type.replace("!", "")
-
-        # Use base mapping
-        return self.GRAPHQL_TO_TS.get(clean_type, clean_type)
-
-    def graphql_to_python(self, graphql_type: str) -> str:
-        """Convert GraphQL type to Python type."""
-        # First convert to TypeScript, then to Python
-        ts_type = self.graphql_to_ts(graphql_type)
-        return self.ts_to_python(ts_type)
-
-    def _is_string_literal(self, ts_type: str) -> bool:
-        """Check if a TypeScript type is a string literal."""
-        ts_type = ts_type.strip()
-        return (ts_type.startswith("'") and ts_type.endswith("'")) or (
-            ts_type.startswith('"') and ts_type.endswith('"')
-        )
-
-    def _handle_union_type(self, ts_type: str, converter_func) -> str:
-        """Handle TypeScript union types (A | B | C)."""
-        parts = [part.strip() for part in ts_type.split("|")]
-
-        # Special case: if it's just Type | null or Type | undefined
-        if len(parts) == 2:
-            if "null" in parts or "undefined" in parts:
-                other_type = parts[0] if parts[1] in ["null", "undefined"] else parts[1]
-                return f"Optional[{converter_func(other_type)}]"
-
-        # Check if all parts are string literals
-        all_string_literals = all(self._is_string_literal(part) for part in parts)
-
-        if all_string_literals:
-            # Use Literal for unions of string literals
-            # Keep the quotes for the literal values
-            literal_values = ", ".join(parts)
-            return f"Literal[{literal_values}]"
-
-        # General union case
-        converted_parts = []
-        for part in parts:
-            if self._is_string_literal(part):
-                # For individual string literals in mixed unions, use Literal
-                converted_parts.append(f"Literal[{part}]")
-            else:
-                converted_parts.append(converter_func(part))
-
-        # Filter out None values for cleaner output
-        converted_parts = [p for p in converted_parts if p != "None"]
-
-        if len(converted_parts) == 1:
-            return f"Optional[{converted_parts[0]}]"
-        elif len(converted_parts) > 1:
-            return f'Union[{", ".join(converted_parts)}]'
-        else:
-            return "None"
+from dipeo.infrastructure.codegen.type_system import (
+    TypeConverter,
+    camel_case,
+    camel_to_snake,
+    kebab_case,
+    pascal_case,
+    pascal_to_camel,
+    snake_case,
+    snake_to_pascal,
+)
 
 
 # ============================================================================
