+++ dipeo/infrastructure/codegen/type_system/__init__.py
@@ -1,0 +1,23 @@
+"""Shared type system utilities for code generation components."""
+
+from .converter import (
+    TypeConverter,
+    camel_case,
+    camel_to_snake,
+    kebab_case,
+    pascal_case,
+    pascal_to_camel,
+    snake_case,
+    snake_to_pascal,
+)
+
+__all__ = [
+    "TypeConverter",
+    "snake_case",
+    "camel_case",
+    "pascal_case",
+    "kebab_case",
+    "camel_to_snake",
+    "snake_to_pascal",
+    "pascal_to_camel",
+]
